+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to shelley are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Merged the `versions` command into `find`. Run `shelley find <tool> -v`
+  (or `--verbose`) to see the full, paginated list of container versions.
+
+### Removed
+
+- `versions` command. Use `shelley find <tool> -v` instead.
+
 ## [0.1.0] - 2026-07-03
 
 First tagged release.